+++ packages/contracts/contracts/test/tokens/SimpleERC1155Token.sol
@@ -4,15 +4,16 @@ pragma solidity ^0.8.17;
 
 import "@openzeppelin/contracts/token/ERC1155/ERC1155.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
+import "./ISimpleToken.sol";
 
-contract SimpleERC1155Token is ERC1155, Ownable {
+contract SimpleERC1155Token is ISimpleERC1155Token, ERC1155, Ownable {
     constructor() ERC1155("Simple") {}
 
     function reserveTokens(
         address account,
         uint256 id,
         uint256 amount
-    ) external {
+    ) external virtual override {
         _mint(account, id, amount, "");
     }
 }