+++ packages/e2e-tests/tests/EndToEnd.ts
@@ -80,6 +80,7 @@ describe("Wallet, Context, Bundler, and SubtreeUpdater", async () => {
 
   beforeEach(async () => {
     docker = new Dockerode();
+    await sleep(5_000);
     hhContainer = await startHardhatNetwork(docker, {
       blockTime: 3_000,
       keys: KEYS,